# src/modules/reports/dto/update-report.dto.ts
import { PartialType, OmitType } from '@nestjs/swagger';
import { CreateReportDto } from './create-report.dto';

export class UpdateReportDto extends PartialType(
  OmitType(CreateReportDto, ['type'] as const),
) {} 