# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/services/{id}/status": {
      "put": {
        "operationId": "updateStatus",
        "summary": "Update service status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service status updated successfully"
          }
        },
        "tags": [
          "Services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/payments/{id}/status": {
      "put": {
        "operationId": "updateStatus",
        "summary": "Update payment status (Admin)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment status updated"
          }
        },
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/reviews": {
      "post": {
        "operationId": "create",
        "summary": "Create new review",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review created successfully"
          }
        },
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "findAll",
        "summary": "Get all reviews",
        "parameters": [
          {
            "name": "service",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return all reviews"
          }
        },
        "tags": [
          "Reviews"
        ]
      }
    },
    "/reviews/{id}": {
      "get": {
        "operationId": "findOne",
        "summary": "Get review details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return review details"
          }
        },
        "tags": [
          "Reviews"
        ]
      },
      "delete": {
        "operationId": "remove",
        "summary": "Delete review (Admin)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review deleted successfully"
          }
        },
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reviews/{id}/provider-response": {
      "put": {
        "operationId": "addProviderResponse",
        "summary": "Add provider response to review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderResponseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider response added successfully"
          }
        },
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reviews/{id}/status": {
      "put": {
        "operationId": "updateStatus",
        "summary": "Update review status (Admin)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review status updated successfully"
          }
        },
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reviews/{id}/helpful": {
      "put": {
        "operationId": "markHelpful",
        "summary": "Mark review as helpful",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review marked as helpful"
          }
        },
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reviews/{id}/report": {
      "put": {
        "operationId": "reportReview",
        "summary": "Report review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review reported successfully"
          }
        },
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/notifications/{id}/status": {
      "put": {
        "operationId": "updateStatus",
        "summary": "Update notification status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification status updated"
          }
        },
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/reports": {
      "post": {
        "operationId": "create",
        "summary": "Create new report",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Report created successfully"
          }
        },
        "tags": [
          "Reports"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "findAll",
        "summary": "Get all reports",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all reports"
          }
        },
        "tags": [
          "Reports"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reports/generate": {
      "post": {
        "operationId": "generate",
        "summary": "Generate quick report",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Report generated successfully"
          }
        },
        "tags": [
          "Reports"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/reports/{id}": {
      "get": {
        "operationId": "findOne",
        "summary": "Get report by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return report details"
          }
        },
        "tags": [
          "Reports"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "update",
        "summary": "Update report",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Report updated successfully"
          }
        },
        "tags": [
          "Reports"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "remove",
        "summary": "Delete report",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report deleted successfully"
          }
        },
        "tags": [
          "Reports"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "HizmetBul API",
    "description": "HizmetBul Servis Platformu API Dokümantasyonu",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdatePaymentStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "completed",
              "failed",
              "refunded"
            ]
          },
          "transactionId": {
            "type": "string",
            "example": "TR123456789"
          },
          "error": {
            "type": "string",
            "example": "Payment failed: Insufficient funds"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
      "CreateReviewDto": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "rating": {
            "type": "number",
            "example": 5,
            "description": "Rating between 1-5"
          },
          "comment": {
            "type": "string",
            "example": "Çok memnun kaldım, kesinlikle tavsiye ederim.",
            "description": "Review comment"
          },
          "images": {
            "example": [
              "image1.jpg",
              "image2.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tags": {
            "example": [
              "profesyonel",
              "dakik",
              "temiz"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "bookingId",
          "rating",
          "comment"
        ]
      },
      "ProviderResponseDto": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string",
            "example": "Değerlendirmeniz için teşekkür ederim.",
            "description": "Provider response to the review"
          }
        },
        "required": [
          "comment"
        ]
      },
      "UpdateReviewStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "UpdateNotificationStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "unread",
              "read",
              "archived"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "FiltersDto": {
        "type": "object",
        "properties": {
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "services": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "providers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "customers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "statuses": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateReportDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Aylık Gelir Raporu"
          },
          "type": {
            "type": "string",
            "enum": [
              "booking",
              "revenue",
              "user",
              "service",
              "performance"
            ]
          },
          "period": {
            "type": "string",
            "enum": [
              "daily",
              "weekly",
              "monthly",
              "yearly",
              "custom"
            ]
          },
          "startDate": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-31"
          },
          "filters": {
            "$ref": "#/components/schemas/FiltersDto"
          },
          "fileType": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          }
        },
        "required": [
          "title",
          "type",
          "period",
          "startDate",
          "endDate"
        ]
      },
      "GenerateReportDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "booking",
              "revenue",
              "user",
              "service",
              "performance"
            ]
          },
          "period": {
            "type": "string",
            "enum": [
              "daily",
              "weekly",
              "monthly",
              "yearly",
              "custom"
            ]
          },
          "format": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "period"
        ]
      },
      "UpdateReportDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Aylık Gelir Raporu"
          },
          "period": {
            "type": "string",
            "enum": [
              "daily",
              "weekly",
              "monthly",
              "yearly",
              "custom"
            ]
          },
          "startDate": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-31"
          },
          "filters": {
            "$ref": "#/components/schemas/FiltersDto"
          },
          "fileType": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          }
        }
      }
    }
  }
}